+++ example/gen/src/SwaggerPetstoreOpenAPI3Client.php
@@ -46,11 +46,9 @@ use Psr\Http\Message\ResponseInterface;
 
 class SwaggerPetstoreOpenAPI3Client
 {
-    /** @var ClientInterface */
-    private $client;
+    private ClientInterface $client;
 
-    /** @var ContainerInterface */
-    private $container;
+    private ContainerInterface $container;
 
     public function __construct(ClientInterface $client, ContainerInterface $container)
     {